Changes based on review
Added a thermally fused MOV to the power input. Added a voltage
reference for the ADC. Fused the input of the capacitive power supply
and the measurement circuit. The thermal protection fuse is monitored,
by powering the AMC1100 unisolated input from the fuse tap. If the input
Added a TVC to the input of the AMC1100 to protect from ESD. Changed the
single ended filter on the AMC1100 to a differential one. Added a
differential filter to the adc input.